@@ -6,23 +6,24 @@
     "@material-ui/core": "^4.11.2",
     "@material-ui/icons": "^4.11.2",
     "@testing-library/jest-dom": "^5.11.4",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
     "@unform/core": "^2.1.6",
     "@unform/web": "^2.1.6",
-    "axios": "^0.21.1",
-    "dotenv": "^10.0.0",
-    "faker": "^5.1.0",
-    "node-sass": "4.14.1",
+    "axios": "^1.3.4",
+    "dotenv": "^16.0.0",
+    "faker": "^6.6.6",
+    "sass": "^1.60.0",
     "prop-types": "^15.7.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-loader-spinner": "^4.0.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.1",
-    "react-toastify": "^7.0.4",
-    "web-vitals": "^0.2.4",
-    "yup": "^0.32.9"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-loader-spinner": "^5.3.4",
+    "react-router-dom": "^6.12.1",
+    "react-scripts": "5.0.1",
+    "react-toastify": "^9.1.3",
+		"sass-loader": "^13.3.2",
+    "web-vitals": "^3.3.2",
+    "yup": "^1.2.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -49,7 +50,7 @@
     ]
   },
   "devDependencies": {
-    "eslint": "^7.16.0",
+    "eslint": "^8.42.0",
     "eslint-plugin-react": "^7.21.5"
   }
 }
