@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { Avatar, IconButton } from '@material-ui/core'
-import { container, btn, avatar } from './navbar.module.scss'
+import styles from './navbar.module.scss'
 
 import { useLoaderContext } from 'contexts/LoaderContext'
 import { useUserContext } from 'contexts/UserContext'
@@ -11,10 +11,10 @@ const Navbar = () => {
 	const { name } = useUserContext()
 
 	return (
-		<header className={container}>
+		<header className={styles.container}>
 			<img src={'/images/jamtastic-white-logo.png'} alt='jamtastic' />
-			<IconButton className={btn} disabled={isLoading}>
-				<Avatar className={avatar}>{name ? name[0].toUpperCase() : null}</Avatar>
+			<IconButton className={styles.btn} disabled={isLoading}>
+				<Avatar className={styles.avatar}>{name ? name[0].toUpperCase() : null}</Avatar>
 			</IconButton>
 		</header>
 	)
