@@ -8,7 +8,7 @@ import TextInput from 'components/commons/atomics/text-input/text-input'
 
 import { login } from 'services/axios/auth-api-calls'
 
-import { container, form, revalidationButtons } from '../forms.module.scss'
+import styles from '../forms.module.scss'
 import { useUserContext } from 'contexts/UserContext'
 import { getStateFromStorage } from 'helpers/storage'
 
@@ -23,14 +23,14 @@ const LoginForm = ({ goToRegister, goToConfirmation, onClose }) => {
 				email: Yup.string().email('Digite um e-mail válido').required('E-mail obrigatório'),
 				password: Yup.string().min(6, 'A senha deve possuir no mínimo 6  caracteres').required('Senha obrigatória')
 			})
-	
+
 			await validate.validate(data, {
 				abortEarly: false
 			})
 
 			await login(data, rememberLogin, () => setName(getStateFromStorage('name')))
 			onClose()
-			
+
 		} catch (error) {
 			if(error instanceof Yup.ValidationError) {
 				const errorMessages = {}
@@ -38,16 +38,16 @@ const LoginForm = ({ goToRegister, goToConfirmation, onClose }) => {
 				error.inner.forEach(err => {
 					errorMessages[err.path] = err.message
 				})
-				
+
 				formRef.current.setErrors(errorMessages)
 			}
 		}
 	}
 
 	return (
-		<div className={container}>
+		<div className={styles.container}>
 			<h1>Mas antes faça seu login!</h1>
-			<Form ref={formRef} className={form} onSubmit={onSubmit}>
+			<Form ref={formRef} className={styles.form} onSubmit={onSubmit}>
 				<TextInput label="Email" type="text" name="email"  />
 				<TextInput label="Senha" type="password" name="password"  />
 				<FormControlLabel label="Lembre-se de mim" control={
@@ -55,11 +55,11 @@ const LoginForm = ({ goToRegister, goToConfirmation, onClose }) => {
 				}/>
 				<Button variant="contained" color="secondary" type="submit">Entrar</Button>
 			</Form>
-			<div className={revalidationButtons}>
+			<div className={styles.revalidationButtons}>
 				<Button variant="text" color="secondary" size="small">Esqueci minha senha</Button>
-				<Button 
-					variant="text" 
-					color="secondary" 
+				<Button
+					variant="text"
+					color="secondary"
 					size="small"
 					onClick={goToConfirmation}
 				>
